% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustMet2.r
\name{adjustMet2}
\alias{adjustMet2}
\title{Phase III sample size adjustment factor (Method 2).}
\usage{
adjustMet2(p2d, p2r, p2e, p2p0 = NULL, p2p1 = NULL, p2a = NULL,
  p2b = NULL, p3p0 = NULL, p3p1 = NULL, p3a = NULL, p3b = NULL,
  nsimul = 5000, seed = NULL, rhorange = c(0.5, 5), p3mpt = 0.001,
  rhot = 1e-04)
}
\arguments{
\item{p2d}{Dataframe with Phase II design, with similar as in \code{\link{EKOptAdaptDesigns}}.}

\item{p2r}{Dataframe containing results of Phase II trials following the design \code{p2d}. It
is the output of the function \code{\link{AnalyzeEKOAD}}.}

\item{p2e}{Phase II estimate to consider among the estimates used by code{\link{AnalyzeEKOAD}}. It
can be \code{"pip"} (naive MLE) or one of the four estimates from methods proposed by
Nhacolo and Brannath (2018): \code{"pim1"}, \code{"pim2"}, \code{"pim2v2"} or \code{"pim3"}.}

\item{p2p0}{Phase II response rate under \eqn{H_0}. If \code{NULL} (default), the value is taken \code{p2d}.}

\item{p2p1}{Phase II response rate under \eqn{H_1}. If \code{NULL} (default), the value is taken \code{p2d}.}

\item{p2a}{Phase II type I error rate. If \code{NULL} (default), the value is taken \code{p2d}.}

\item{p2b}{Phase II type II error rate. If \code{NULL} (default), the value is taken \code{p2d}.}

\item{p3p0}{Phase III response rate of the control group. If \code{NULL} (default), the value is set to \code{p2p0}.}

\item{p3p1}{Phase III response rate of the treatment group. If \code{NULL} (default), the value is set to \code{p2p1}.}

\item{p3a}{Phase III type I error rate. If \code{NULL} (default), the value is set \code{p2a}.}

\item{p3b}{hase III type II error rate. If \code{NULL} (default), the value is set \code{p2b}.}

\item{nsimul}{Number of (parametric) bootstrap samples (default 5000).}

\item{seed}{Seed for random number generator. If \code{NULL} (default), no seed is set.}

\item{rhorange}{A vector specifying a range to serach for \eqn{\rho}. The default is \code{c(0.5,5)}.}

\item{p3mpt}{Tolerated error margin for the power, i.e., maximum allowed absolute diference between the estimated
expected power and the target. The default is 0.001.}

\item{rhot}{Search for \eqn{\rho} is interrupted and deem unsuccessful if the absolute difference between
current and the previous is less than or equal to \code{rhot}.}
}
\value{
A list containing two dataframes \code{final} and \code{intermed}. \code{final} contains the final
measures for the adjustment factor (\eqn{\rho}), and for the unadjusted and adjusted power. \code{intermed}
holds the intermediate results (of each bootstrap sample).
}
\description{
\code{adjustMet2} calculates the multiplicative ajustment  factor \eqn{\rho} to be applied to Phase III
sample size estimate using Method 2 proposed by Nhacolo and Brannath (in press).
}
\details{
The aim of the adjustment is to get an adequately powered Phase III trial based
on Phase II data. \eqn{\rho} is found using numerical rearch. See the documentation of the function
\code{\link{AnIItoIIIRe}} for more details about the designs.
}
\examples{
\dontrun{
vdid <- c(6,10) # design ids
vp2est <- c("pip","pim1","pim2","pim2v2","pim3")
nse <- 1000#number of simulations for each phase
cur <- 1; tot <- length(vdid)*length(vp2est)
for (did in vdid){
 for (p2est in vp2est){
   cat('Processing ',cur,' of ',tot,' (',100*round(cur/tot,1),'\%)\\n',sep = '')
    load(paste0("p2r",did,".rdata")) # output of the function AnalyzeEKOAD
    out <- adjustMet2(p2d = EKOADwn[EKOADwn$id==did,], p2r = rslt[1:nse,], p2e = p2est, nsimul = nse, seed = 3343)
    write.csv(out$final,file = paste0("final",did,p2est,".csv"),row.names = FALSE)
   write.csv(out$intermed,file = paste0("intermed",did,p2est,".csv"),row.names = FALSE)
   cur <- cur+1
 }
}


vdid <- c(6,10)
vp2est <- c("pip","pim1","pim2","pim2v2","pim3")
fa <- data.frame()
for (did in vdid)
{
 for (p2est in vp2est){
    f <- read.csv(paste0("final",did,p2est,".csv"))
    fn <- names(f)
    f$dsgn <- did
    f <- f[,c('dsgn',fn)]
    fa <- rbind(fa,f)
  }
}
write.csv(fa,file = "final_all.csv",row.names = FALSE)
}
}
\references{
Nhacolo, A. and Brannath, W. Using Estimates from Adaptive Phase II Oncology Trials to Plan Phase III Trials.
\emph{In press}.

Nhacolo, A. and Brannath, W. Interval and point estimation in adaptive Phase II trials with binary endpoint.
\emph{Stat Methods Med Res}, 2018.

Ahn, C., Heo, M. and Zhang, S. \emph{Sample Size Calculations for Clustered and Longitudinal Outcomes in Clinical Research}.
CRC Press, 2014.
}
\seealso{
\code{\link{adjustMet2}}, \code{\link{SimulateEKOAD}}, \code{\link{AnalyzeEKOAD}}.
}
\author{
Arsenio Nhacolo
}
