% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyzePerformanceSimon.r
\name{AnalyzePerformanceSimon}
\alias{AnalyzePerformanceSimon}
\title{Performance of estimation methods}
\usage{
AnalyzePerformanceSimon(designs = "all", basedir = NA)
}
\arguments{
\item{designs}{Taking values \code{"fixed"}, \code{"adaptive"} or \code{"all"},
indicating whether only classical, adaptive or all designs should be included.
The default is \code{"all"}.}

\item{basedir}{The root directory in which simulations were performed. The current
working directory is assumed by default. It must contain all the files and folders
created by \code{\link{SimulateSimonDsgn}} and/or \code{\link{SimulateSimonDsgnAdaptN}}.}
}
\value{
Dataframe containing bias, mean square error and variance of the estimators,
power, expected sample size, and design information.
}
\description{
It takes the results produced by \code{\link{AnalyzeSimonDsgn}} and
\code{\link{AnalyzeSimonDsgnAdaptN}} and produces a dataframe containing bias,
mean square error and variance of the estimators. It also calculates the power
and the expected sample size (EN) where applicable.
}
\details{
Computations are done for different combinations of values of
\code{stop}, (0,1), and \code{success}, (0,1). See \code{\link{AnalyzeSimonDsgn}} or
\code{\link{AnalyzeSimonDsgnAdaptN}}. For instance, computations done on all
simulated trials are marked with \code{"both"} in the columns \code{stop} and
\code{success}, while the ones done only on trials that continued to the final
stage have \code{stop = "no"} and \code{success = "both"}.
}
\examples{
\dontrun{
AnalyzePerformanceSimon()
}
}
\seealso{
\code{\link{AnalyzeSimonDsgn}}, \code{\link{AnalyzeSimonDsgnAdaptN}},
\code{\link{pdata}} and \code{\link{AnalyzePerformanceSimon2}}.
}
\author{
Arsenio Nhacolo
}
