% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyzeSimonDsgn.r
\name{AnalyzeSimonDsgn}
\alias{AnalyzeSimonDsgn}
\title{Analysis of simulated Simon's design trials}
\usage{
AnalyzeSimonDsgn(replicates = NA, basedir = NA)
}
\arguments{
\item{replicates}{Number of trials to be analysed. By default all simulated trials are
analysed.}

\item{basedir}{The root directory in which simulations were performed. The current
working directory is assumed by default. It must contain all the files and folders
created by \code{\link{SimulateSimonDsgn}}.}
}
\value{
Creates two data files in \code{basedir} containing results for optimal
(\emph{ResultsOptimalDesign.csv}) and minimax (\emph{ResultsMinimaxDesign.csv}).
The files contain a \code{trial} ID, stage 1, stage 2 and overall number of
successful responses, \code{s1}, \code{s2} and \code{s}, sample sizes
(equal to those pre-specified by design),  \code{n1}, \code{n2} and \code{n},
and critical values, \code{r1} and \code{r}. \code{p0} the response rate
assumed under \eqn{H_0} and \code{dsgnp1} under \eqn{H_1}. \code{p1} is the
true response rate (used for generating trial data). \code{pm1} and \code{pm2}
are, respectively, \code{\link{pm}} based only of stage 1 and stage 2 data.
\code{stop} indicates whether the trial stopped at first stage (\code{stop = 1}),
and \code{success} indicates whether \eqn{H_0} was rejected (\code{success = 1}).
}
\description{
Analyses the trials simulated by \code{\link{SimulateSimonDsgn}}.
}
\details{
In addition to hypothesis testing, the response rate is estimated
using different estimators: \code{\link{pm}}, \code{\link{pg}}, \code{\link{pu}},
\code{\link{pp}} and \code{\link{pk}}.
}
\examples{
AnalyzeSimonDsgn()
}
\seealso{
\code{\link{CalculateSimonDsgn}}, \code{\link{SimulateSimonDsgn}},
\code{\link{AnalyzePerformanceSimon}} and \code{\link{AnalyzeSimonDsgnAdaptN}}.
}
\author{
Arsenio Nhacolo
}
