% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnIItoIIIRe.r
\name{AnIItoIIIRe}
\alias{AnIItoIIIRe}
\title{Use of Phase II estimates to plan Phase III sample size.}
\usage{
AnIItoIIIRe(rslt, f = c(0.95, 0.96, 0.97, 0.98, 0.99))
}
\arguments{
\item{rslt}{Dataframe containing the output from the function \code{\link{AnalyzeEKOAD}}, but with
only successful trials (rslt$suco==1), i.e., trials in which \eqn{H_0} was rejected.}

\item{f}{Vector of length 5 containing multiplicative ajustment factors to be applied
to Phase II estimates. The default is \code{f = c(.95,.96,.97,.98,.99)}.}
}
\value{
The input dataframe with corresponding Phase III sample size and power.
}
\description{
\code{AnIItoIIIRe} calculates the power in a Phase III equal-size group
two-arm randomized clinical trial with a binary response planned using estimates
from Phase II adaptive two-stage trial.
}
\details{
The sample size (N) of the Phase III trial is based on the estimates
naive MLE and estimators proposed by Nhacolo and Brannath (2018). Different values of retention
factor f proposed by Kirby et al. (2012) are applied.The control group response rate is considered
to be equal to that under the null hypothesis of the Phase II design, and the hypothesized treatment
group response rate considered to be equal to that estimated from the Phase II trial. The target type I
error and power are the same as of the Phase II design.Two-sided hypothesis test is assumed.N is a sample
size per group, and equal size groups are assume. Hence, N total is 2*N.
When calculating the power, the true response rate (in treatment group) is considered to be the one under
which the Phase II trial was simulated (spi1).
}
\references{
Nhacolo, A. and Brannath, W. Using Estimates from Adaptive Phase II Oncology Trials to Plan Phase III Trials.
\emph{In press}.

Nhacolo, A. and Brannath, W. Interval and point estimation in adaptive Phase II trials with binary endpoint.
\emph{Stat Methods Med Res}, 2018.

Ahn, C., Heo, M. and Zhang, S. \emph{Sample Size Calculations for Clustered and Longitudinal Outcomes in Clinical Research}.
CRC Press, 2014.
}
\seealso{
\code{\link{AnalyzeEKOAD}}, \code{\link{SimulateEKOAD}}, \code{\link{PerforIItoIIIRe}}.
}
\author{
Arsenio Nhacolo
}
