% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aop1.r
\name{aop1}
\alias{aop1}
\title{Overall p-value (Method 1 of Nhacolo and Brannath, 2018).}
\usage{
aop1(dsgn, x1o, xo, verbose = TRUE)
}
\arguments{
\item{dsgn}{Dataframe containing one of the designs in \code{\link{EKOADwn}}.}

\item{x1o}{The observed stage 1 number of responses.}

\item{xo}{The total observed number of responses.}

\item{verbose}{If \code{TRUE} (default) messages will be printed.}
}
\value{
p-value.
}
\description{
\code{aop1} calculates the overall p-value for adaptive two-stage designs
with binary endpoint using the Method 1 (see Nhacolo and Brannath, 2018).
}
\details{
This is one of the four methods proposed by Nhacolo and Brannath (2018) primarily
for single-arm adaptive two-stage group sequential designs with a binary endpoint.
}
\references{
Nhacolo, A. and Brannath, W. Interval and point estimation in adaptive Phase II trials with binary endpoint.
\emph{Stat Methods Med Res}, 2018.
}
\seealso{
\code{\link{aop2}}, \code{\link{aop2v2}}, \code{\link{aop3e}}.
}
\author{
Arsenio Nhacolo
}
