% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aop1e.r
\name{aop1e}
\alias{aop1e}
\title{Overall p-value for CI (Method 1 of Nhacolo and Brannath, 2018).}
\usage{
aop1e(dsgn, x1o, xo, newpi0)
}
\arguments{
\item{dsgn}{Dataframe containing one of the designs in \code{\link{EKOADwn}}.}

\item{x1o}{The observed stage 1 number of responses.}

\item{xo}{The total observed number of responses.}

\item{newpi0}{New response probability that replaces the one under the null hypothesis.}
}
\value{
p-value.
}
\description{
\code{aop1e} is a modified version of \code{\link{aop1}} used
for getting the confidence interval.
}
\details{
This is one of the four methods proposed by Nhacolo and Brannath (2018) primarily
for single-arm adaptive two-stage group sequential designs with a binary endpoint.
}
\references{
Nhacolo, A. and Brannath, W. Interval and point estimation in adaptive Phase II trials with binary endpoint.
\emph{Stat Methods Med Res}, 2018.
}
\seealso{
\code{\link{aop2e}}, \code{\link{aop2ev2}}, \code{\link{aop3e}}, \code{\link{aop1}}.
}
\author{
Arsenio Nhacolo
}
