% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aop3e.r
\name{aop3e}
\alias{aop3e}
\title{Overall p-value (Method 3 of Nhacolo and Brannath, 2018).}
\usage{
aop3e(dsgn, x1o, xo, newpi0 = NULL)
}
\arguments{
\item{dsgn}{Dataframe containing one of the designs in \code{\link{EKOADwn}}.}

\item{x1o}{The observed stage 1 number of responses.}

\item{xo}{The total observed number of responses.}

\item{newpi0}{New response probability that replaces the one under the null hypothesis.
Omit it if the intention is only to calculate the overall p-value.}
}
\value{
p-value.
}
\description{
\code{aop3e} calculates the overall p-value for adaptive two-stage designs
with binary endpoint using the Method 3 (see Nhacolo and Brannath, 2018).
}
\details{
This is one of the four methods proposed by Nhacolo and Brannath (2018) primarily
for single-arm adaptive two-stage group sequential designs with a binary endpoint.
}
\references{
Nhacolo, A. and Brannath, W. Interval and point estimation in adaptive Phase II trials with binary endpoint.
\emph{Stat Methods Med Res}, 2018.
}
\seealso{
\code{\link{aop1}}, \code{\link{aop1e}}, \code{\link{aop2}}, \code{\link{aop2e}}, \code{\link{aop2v2}}, \code{\link{aop2ev2}}
}
\author{
Arsenio Nhacolo
}
