% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateSimonDsgn.r
\name{CalculateSimonDsgn}
\alias{CalculateSimonDsgn}
\title{Simon's designs}
\usage{
CalculateSimonDsgn(p0, p1, alpha, beta, verbose = TRUE)
}
\arguments{
\item{p0}{The response rate under the null hypothesis.}

\item{p1}{The response rate under the alternative hypothesis.}

\item{alpha}{Type I error rate.}

\item{beta}{Type II error rate.}

\item{verbose}{If TRUE (default) the designs are printed (gives messy
printout when the function is run without assigment).}
}
\value{
A two-row dataframe containing the optimal and the minimax designs.
}
\description{
\code{CalculateSimonDsgn} finds Simon's optimal and minimax designs.
}
\details{
Simon's designs are two-stage single-arm for phase II clinical trials.
They consist in first stage and overall sample sizes and critical values,
\code{n1} and \code{n}, and \code{r1} and \code{r}, respectively.
}
\examples{
d <- CalculateSimonDsgn(0.2, 0.4, 0.05, 0.1)
}
\references{
Simon, R. Optimal two-stage designs for phase II clinical trials.
\emph{Control Clin Trials}, 1989, 10, 1-10.
}
\seealso{
\code{\link{SimulateSimonDsgn}} and \code{\link{SimulateSimonDsgnAdaptN}}.
}
\author{
Arsenio Nhacolo
}
