% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkMonoDCF.r
\name{checkMonoDCF}
\alias{checkMonoDCF}
\title{Check the monotonicity of the sample space ordering.}
\usage{
checkMonoDCF(d, verbose = TRUE)
}
\arguments{
\item{d}{Dataframe containing one of the designs in \code{\link{EKOADwn}}.}

\item{verbose}{If \code{TRUE} (default) messages about monotonicity will be printed.}
}
\value{
A list containing a dataframe (\code{mono}) with detailed info, and a logical variable
\code{notmono} indicating wether non-monotonicity was concluded.
}
\description{
\code{checkMonoDCF} checks the monotonicity of the sample space ordering defined
based on inverse normal combination function (see Nhacolo and Brannath, 2018).
}
\details{
The monotonicity is with respect to the stage 2 number of successes.
}
\examples{
\dontrun{
#Check for all Englert and Kieser designs
notmov <- c()
for (i in 1:max(EKOADwn$id)){
  notmov <- c(notmov,checkMonoDCF(EKOADwn[EKOADwn$id==1,],verbose=FALSE)[[2]])
}
isMonotone <- !any(notmov);isMonotone
}
}
\references{
Nhacolo, A. and Brannath, W. Interval and point estimation in adaptive Phase II trials with binary endpoint.
\emph{Stat Methods Med Res}, 2018.
}
\author{
Arsenio Nhacolo
}
