% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci2.r
\name{ci2}
\alias{ci2}
\title{Confidence interval (using Method 2 of Nhacolo and Brannath, 2018).}
\usage{
ci2(dsgn, x1o, xo, alpha = 0.05, twosided = FALSE)
}
\arguments{
\item{dsgn}{Dataframe containing one of the designs in \code{\link{EKOADwn}}.}

\item{x1o}{The observed stage 1 number of responses.}

\item{xo}{The total observed number of responses.}

\item{alpha}{The significance level.}

\item{twosided}{If \code{FALSE} (default) a one-sided CI is produced.}
}
\value{
CI is a list with lower and upper bounds.
}
\description{
\code{ci2} computes confidence interval.
}
\details{
This CI is obtained using the Method 2, one of the four methods proposed by
Nhacolo and Brannath (2018) primarily for single-arm adaptive two-stage group sequential
designs with a binary endpoint.
}
\references{
Nhacolo, A. and Brannath, W. Interval and point estimation in adaptive Phase II trials with binary endpoint.
\emph{Stat Methods Med Res}, 2018.
}
\seealso{
\code{\link{ci1}}, \code{\link{ci2v2}}, \code{\link{ci3}}, \code{\link{aop2}}, \code{\link{aop2e}}, \code{\link{pipv2}},
 \code{\link{mue2}}.
}
\author{
Arsenio Nhacolo
}
