% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsgnPrep.r
\name{dsgnPrep}
\alias{dsgnPrep}
\title{Pre-process the Englert and Kieser (2013) optimal adaptive designs.}
\usage{
dsgnPrep(dsgn = NULL, w1 = "n", w2 = NULL)
}
\arguments{
\item{dsgn}{Dataframe containing one of the designs in \code{\link{EKOptAdaptDesigns}}.}

\item{w1, w2}{Stage 1 and 2 weights. If \code{w1="n"} (default), weights a calculated
based on stage-wise sample sizes as described in Nhacolo and Brannath (2018).
If \code{w1="sr2"}, then \code{w1=w2=1/sqrt(2)}.}
}
\value{
Dataframe containing the input dataframe with added information.
}
\description{
\code{dsgnPrep} takes Englert and Kieser's optimal adaptive design and adds
information that is needed by other functions.
}
\details{
The function adds, to each x1 leading to 2nd stage, the corresponding p-value (p1)
and its back-wards image (p1B), the stage-wise weights w1 and w2 and other information used
in inference methods proposed by  Nhacolo and Brannath (2018).
}
\examples{
\dontrun{
#Designs with w1a and w2 calculated based on sample sizes
EKOADwn <- data.frame()
for (j in 1:max(EKOptAdaptDesigns$id)){
  EKOADwn <- rbind(EKOADwn, dsgnPrep(dsgn = EKOptAdaptDesigns[EKOptAdaptDesigns$id==j,],w1 = "n"))
}
save(EKOADwn,file = "EKOADwn.RData")
}
}
\references{
Nhacolo, A. and Brannath, W. Interval and point estimation in adaptive Phase II trials with binary endpoint.
\emph{Stat Methods Med Res}, 2018.
}
\author{
Arsenio Nhacolo
}
