% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EKOADwn.r
\docType{data}
\name{EKOADwn}
\alias{EKOADwn}
\title{Pre-processed Englert and Kieser (2013)'s optimal adaptive designs.}
\format{A dataframe with 709 rows and 20 variables.}
\usage{
EKOADwn
}
\description{
A dataframe containing all the designd in \code{\link{EKOptAdaptDesigns}}
pre-processed by the function \code{\link{dsgnPrep}}, the the argument
\code{w1} set to "n".
}
\keyword{datasets}
