% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN2v2.r
\name{getN2v2}
\alias{getN2v2}
\title{Number of patients to be enrolled in the second stage}
\usage{
getN2v2(cp, p1, design, k, mode = 0, alpha = 0.05)
}
\arguments{
\item{cp}{conditional power to which the number of patients for the second stage should be adjusted.}

\item{p1}{response probability under the alternative hypothesis.}

\item{design}{a dataframe containing all critical values for a Simon's two-stage design defined by the colums r1, n1, r, n and p0.
\itemize{
   \item r1 = critical value for the first stage (more than r1 responses needed to proceed to the second stage).
   \item n1 = number of patients enrolled in the first stage.
   \item r = critical value for the whole trial (more than r responses needed at the end of the study to reject the null hypothesis).
   \item n = number of patients enrolled in the whole trial.
   \item p0 = response probability under the null hypothesis.
}}

\item{k}{number of responses observed at the interim analysis.}

\item{mode}{a value out of {0,1,2,3} dedicating the methode spending the "rest alpha" (difference between nominal alpha level and actual alpha level for the given design).
\itemize{
   \item 0 = "rest alpha" is not used.
   \item 1 = "rest alpha" is spent proportionally.
   \item 2 = "rest alpha" is spent equally.
   \item 3 = "rest alpha" is spent only to the worst case scenario (minimal number of responses at the interim analysis so that the study can proceed to the second stage).
}}

\item{alpha}{overall significance level the trial was planned for.}
}
\description{
Calculates the number of patients which should be enrolled in the
second stage if the conditional power should be altert to "cp". It's a version
of \code{\link[OneArmPhaseTwoStudy]{getN2}}.
}
\details{
This functon is the same as \code{\link[OneArmPhaseTwoStudy]{getN2}}
(\code{OneArmPhaseTwoStudy} package), with some changes in arguments' validation.
It's is a helper to
\code{\link{SimulateSimonDsgnAdaptN}}.
}
\examples{
designParam <- CalculateSimonDsgn(0.2, 0.4, 0.05, 0.1)
dsgn <- designParam[designParam$Type == "Optimal",]
getN2v2(0.9, dsgn$p1, dsgn, 7)
}
\references{
Englert S., Kieser M. Adaptive designs for single-arm phase II trials in oncology.
\emph{Pharm Stat}, 2012, 11, 241-249.
}
\seealso{
\code{\link[OneArmPhaseTwoStudy]{getN2}}, \code{\link{SimulateSimonDsgnAdaptN}}.
}
