% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mue1.r
\name{mue1}
\alias{mue1}
\title{Median estimate (using Method 1 of Nhacolo and Brannath, 2018).}
\usage{
mue1(dsgn, x1o, xo)
}
\arguments{
\item{dsgn}{Dataframe containing one of the designs in \code{\link{EKOADwn}}.}

\item{x1o}{The observed stage 1 number of responses.}

\item{xo}{The total observed number of responses.}
}
\value{
Median estimate of response probability.
}
\description{
\code{mue1} calculates the median estimate of the response rate.
}
\details{
This estimate is obtained using the Method 1, one of the four methods proposed by
Nhacolo and Brannath (2018) primarily for single-arm adaptive two-stage group sequential
designs with a binary endpoint.
}
\references{
Nhacolo, A. and Brannath, W. Interval and point estimation in adaptive Phase II trials with binary endpoint.
\emph{Stat Methods Med Res}, 2018.
}
\seealso{
\code{\link{mue2}}, \code{\link{mue2v2}}, \code{\link{mue3}}, \code{\link{aop1}}, \code{\link{aop1e}}, \code{\link{pipv1}}.
}
\author{
Arsenio Nhacolo
}
