% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nct.r
\name{Nct}
\alias{Nct}
\title{Sample size per group for single-stage parallel-group RCT.}
\usage{
Nct(pc, pt, alp = 0.05, pow = 0.8)
}
\arguments{
\item{pc}{Response probability in control group.}

\item{pt}{Response probability in treatment group.}

\item{alp}{Significance level (default: 0.05).}

\item{pow}{Power (default: 0.8)}
}
\value{
Sample size for one group.
}
\description{
\code{Nct} calculates sample size for one group in an equal-size group
two-arm randomized clinical trial with a binary response.
}
\details{
The sample size is for one group (arm), double the number to get the total.
}
\examples{
Nct(0.2,0.3,0.05,0.9)
}
\references{
Ahn, C., Heo, M. and Zhang, S. \emph{Sample Size Calculations for Clustered and Longitudinal Outcomes in Clinical Research}.
CRC Press, 2014.
}
\seealso{
\code{\link{Pwr}}.
}
\author{
Arsenio Nhacolo
}
