% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdata.r
\name{pdata}
\alias{pdata}
\title{Helper function for analyzing the performance of estimators}
\usage{
pdata(t, design, stop, success, replicates)
}
\arguments{
\item{t}{Dataframe containing results produced by \code{\link{AnalyzeSimonDsgn}}
or \code{\link{AnalyzeSimonDsgnAdaptN}}.}

\item{stop}{Taking value \code{"yes"}, \code{"no"} or \code{"both"}, indicating
that only trials that stopped, continued or both were analyzed.}

\item{success}{Taking value \code{"yes"}, \code{"no"} or \code{"both"}, indicating
that only trials that were successful, unsuccessful or both were analyzed.}

\item{replicates}{Number of trials analysed. It is equal to the number of
rows in \code{t}.}
}
\value{
Dataframe containing bias, mean square error and variance of the estimators.
}
\description{
It takes the results produced by \code{\link{AnalyzeSimonDsgn}} or
\code{\link{AnalyzeSimonDsgnAdaptN}} and produces a dataframe containing bias,
mean square error and variance of the estimators.
}
\details{
It is a helper function for \code{\link{AnalyzePerformanceSimon}}.
It also calculates the power and the expected sample size (EN) where applicable.
}
\examples{
\dontrun{
rslt <- read.csv("ResultsOptimalDesign.csv")
nrep <- nrow(rslt)
t <- rslt
presult <- pdata(t, "Optimal", "both", "both", nrep)
t <- rslt[rslt$stop == 0,]
presult <- rbind(presult, pdata(t, "Optimal", "no", "both", nrep))
}
}
\seealso{
\code{\link{AnalyzeSimonDsgn}}, \code{\link{AnalyzeSimonDsgnAdaptN}} and
\code{\link{AnalyzePerformanceSimon}}.
}
\author{
Arsenio Nhacolo
}
