% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdata2.r
\name{pdata2}
\alias{pdata2}
\title{Helper function for analyzing the performance of estimators}
\usage{
pdata2(t, stop, success, replicates)
}
\value{
Dataframe
}
\description{
\code{pdata2} is a helper function used by function \code{\link{PerformanceEKOAD}}.
}
\details{
Not to be used directly.
}
\author{
Arsenio Nhacolo
}
