% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerforIItoIIIRe.r
\name{PerforIItoIIIRe}
\alias{PerforIItoIIIRe}
\title{Performance, with respect to Phase III power, of phase II estimates.}
\usage{
PerforIItoIIIRe(t)
}
\arguments{
\item{t}{Dataframe containing the output from the function \code{\link{AnIItoIIIRe}}.}
}
\value{
Dataframe.
}
\description{
\code{PerforIItoIIIRe} calculates the mean and median power in a Phase III
trials from the output of \code{\link{AnIItoIIIRe}}.
}
\examples{
\dontrun{
rslt <- read.csv("ResultsAll.csv")
rsltfull <- rslt
rslt <- rslt[rslt$suco==1,]
rslt <- rslt[,c("pi0", "pi1", "spi1", "alpha", "beta", "suco", "pip",
                "pim1", "pim2", "pim2v2", "pim3")]
rslt <- rslt[rslt$spi1>=rslt$pi0+0.1 & rslt$spi1<=rslt$pi1+0.3,]
rslt$spi1f <- factor(rslt$spi1)
cats <- levels(rslt$spi1f)
ncats <- length(cats)
setwd(paste0("C:/Users/arsenio/Documents/PhD/Simulations/Paper2/Reuse/pi01by0.01/50000/",did))
save(ncats,file = "ncats.rdata")
for (i in 1:ncats){
  sr <- rslt[rslt$spi1f==cats[i],]#Single result (result of a specific spi1)
  save(sr,file = paste0("sr",i,".rdata"))
}
load("ncats.rdata")
PerfAll <- data.frame()
for (k in 1:ncats){
 load(paste0("sr",k,".rdata"))
 sre <- AnIItoIIIRe(rslt = sr,f = c(.95,.96,.97,.98,.99))
  PerfAll <- rbind(PerfAll,PerforIItoIIIRe(sre))
  rm(sr)
}
write.csv(PerfAll, file = "PerfAllIItoIII.csv", row.names = F)
}
}
\references{
Nhacolo, A. and Brannath, W. Using Estimates from Adaptive Phase II Oncology Trials to Plan Phase III Trials.
\emph{In press}.
}
\seealso{
\code{\link{AnIItoIIIRe}}.
}
\author{
Arsenio Nhacolo
}
