% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerformanceEKOAD.r
\name{PerformanceEKOAD}
\alias{PerformanceEKOAD}
\title{Performance of estimation methods}
\usage{
PerformanceEKOAD(basedir = NULL)
}
\arguments{
\item{basedir}{The base directory containing the file with the results (\code{Results.csv}).
If \code{NULL} (default), the current working directory is uded.}
}
\value{
A dataframe with the performance results. A copy is saved in the file the
 \code{PerformanceResults.csv} in the \code{basedir}.
}
\description{
\code{PerformanceEKOAD} calculates performance measures (bias, mean
square error, coverage probability) of the estimation methods based on the
results produced by \code{\link{AnalyzeEKOAD}}.
}
\examples{
\dontrun{
#SIMULATIONS
for (did in c(6,10)){#Design ID
cat("============================ Design ",did," ============================\\n")
repl <- 50000 # number of replicated trials for each p
dir.create(as.character(did))
setwd(as.character(did))
design <- EKOADwn[EKOADwn$id==did,]
seed = 3343
if (file.exists("PerforAll.csv")) unlink("PerforAll.csv")
piv <- seq(0,1,0.025) # p to simulate data
resul <- data.frame()
perf <- data.frame()
k <- 0
pl <- length(piv)
for (pi in piv){
  k <- k+1
  cat("_________________________ pi = ",pi, " (",k," of ",pl,") _________________________\\n",sep = "")
  SimulateEKOAD(replicates = repl, dsgn = design, newpi1 = pi, seed = seed)
  resul <- rbind(resul, AnalyzeEKOAD())
  perf <- rbind(perf, PerformanceEKOAD())
}
write.table(resul, file ="ResultsAll.csv", sep = ",", row.names = F, col.names = TRUE)
write.table(perf, file ="PerforAll.csv", sep = ",", row.names = F, col.names = TRUE)
cat("Design ID: ", design$id[1], "\\nReplicates: ", repl, "\\nSeed: ", seed,
    "\\nDate last run: ", date(),file = "info.txt", sep = "", append = FALSE)
}
}
}
\references{
Nhacolo, A. and Brannath, W. Interval and point estimation in adaptive Phase II trials with binary endpoint.
\emph{Stat Methods Med Res}, 2018.
}
\seealso{
\code{\link{SimulateEKOAD}}, \code{\link{AnalyzeEKOAD}}.
}
\author{
Arsenio Nhacolo
}
