% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg.r
\name{pg}
\alias{pg}
\title{Bias-reduced  estimator}
\usage{
pg(s, n1, r1, n)
}
\arguments{
\item{s}{Total number of successes.}

\item{n1}{Stage 1 sample size.}

\item{r1}{Stage 1 critical value (trial is stopped at stage 1 if the number of successes
is at most \code{r1}).}

\item{n}{Total sample size.}
}
\value{
Estimate of the response rate.
}
\description{
Calculates the bias-reduced estimator of the true
response rate as proposed by \emph{Guo and Liu (2005)}.
}
\details{
It uses bias subtraction, with bias calculated by \code{\link{sbias}}
and response rate estimated by \code{\link{pm}}.
}
\examples{
pg(21, 19, 4, 54)
}
\references{
Guo, H. Y. and Liu, A. A simple and efficient bias-reduced estimator of
response probability following a group sequential phase II trial.
\emph{J Biopharm Stat},   2005, 15, 773-781.
}
\seealso{
\code{\link{sbias}}, \code{\link{pm}}, \code{\link{pu}}, \code{\link{pp}} and
\code{\link{pk}}.
}
\author{
Arsenio Nhacolo
}
