% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipv1.r
\name{pipv1}
\alias{pipv1}
\title{Response rate to attain a specified p-value (using Method 1 of Nhacolo and Brannath, 2018).}
\usage{
pipv1(dsgn, x1o, xo, pv)
}
\arguments{
\item{dsgn}{Dataframe containing one of the designs in \code{\link{EKOADwn}}.}

\item{x1o}{The observed stage 1 number of responses.}

\item{xo}{The total observed number of responses.}

\item{pv}{The desired p-value.}
}
\value{
Response probability.
}
\description{
\code{pipv1} finds the response probability under the null hypotheisis that,
given the observed data, would yield a desired overall p-value.
}
\details{
The p-value is obtained using the Method 1, one of the four methods proposed by
Nhacolo and Brannath (2018) primarily for single-arm adaptive two-stage group sequential
designs with a binary endpoint.
}
\references{
Nhacolo, A. and Brannath, W. Interval and point estimation in adaptive Phase II trials with binary endpoint.
\emph{Stat Methods Med Res}, 2018.
}
\seealso{
\code{\link{pipv2}}, \code{\link{pipv2v2}}, \code{\link{pipv3}}, \code{\link{aop1}}, \code{\link{aop1e}}.
}
\author{
Arsenio Nhacolo
}
