% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk.r
\name{pk}
\alias{pk}
\title{Median unbiased estimator}
\usage{
pk(s, n1, r1, n, p0)
}
\arguments{
\item{s}{Total number of successes.}

\item{n1}{Stage 1 sample size.}

\item{r1}{Stage 1 critical value (trial is stopped at stage 1 if the number of successes
is at most \code{r1}).}

\item{n}{Total sample size.}

\item{p0}{Response rate under the null hypothesis.}
}
\value{
Estimate of the response rate.
}
\description{
Calculates the median unbiased estimator of true response rate
 for for Simon-like designs.
}
\details{
Median unbiased estimator is the value of response rate such that
the p-value is 0.5 (\emph{Koyama and Chen, 2008}). The solution is found using
numerical search, with a precision of 0.000001.
}
\examples{
pk(21, 19, 4, 54, 0.2)
}
\references{
Koyama, T. and Chen, H. Proper inference from Simon's two-stage designs.
\emph{Stat Med}, 2008, 27, 3145-3154.
}
\seealso{
\code{\link{pvaluek}}, \code{\link{pquantile}}, \code{\link{pm}},
\code{\link{pg}}, \code{\link{pu}} and \code{\link{pp}}.
}
\author{
Arsenio Nhacolo
}
