% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm.r
\name{pm}
\alias{pm}
\title{Sample proportion}
\usage{
pm(s, n)
}
\arguments{
\item{s}{Total number of successes.}

\item{n}{Total sample size.}
}
\value{
Estimate of the response rate.
}
\description{
Calculates the sample proportion.
}
\details{
For fixed designs the sample propotion is an unbiased (maximum likelihood)
estimator of the response rate, but in group sequential designs (e.g., Simon's)
it is biased.
}
\examples{
pm(21, 54)
}
\seealso{
\code{\link{pg}}, \code{\link{pu}}, \code{\link{pp}} and \code{\link{pk}}.
}
\author{
Arsenio Nhacolo
}
