% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp.r
\name{pp}
\alias{pp}
\title{UMVCUE}
\usage{
pp(s, n1, r1, n)
}
\arguments{
\item{s}{Total number of successes.}

\item{n1}{Stage 1 sample size.}

\item{r1}{Stage 1 critical value (trial is stopped at stage 1 if the number of successes
is at most \code{r1}).}

\item{n}{Total sample size.}
}
\value{
Estimate of the response rate.
}
\description{
Calculates the uniformly minimum variance conditionally unbiased estimator
(UMVCUE) of the true response probability.
}
\details{
The UMVCUE (\emph{Pepe et al., 2009}) is conditional on on proceeding to the
second stage. he sample proportion is used when the trial stopped at first stage.
}
\examples{
pp(21, 19, 4, 54)
}
\references{
Pepe, M. S.; Feng, Z.; Longton, G. and Koopmeiners, J. Conditional estimation
of sensitivity and specificity from a phase 2 biomarker study allowing early termination
for futility. \emph{Stat Med}, 2009, 28, 762-779.
}
\seealso{
\code{\link{pm}}, \code{\link{pg}}, \code{\link{pu}} and
\code{\link{pk}}.
}
\author{
Arsenio Nhacolo
}
