% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pquantile.r
\name{pquantile}
\alias{pquantile}
\title{Value of response rate to attain a given p-value}
\usage{
pquantile(s, n1, r1, n, p0, pvalue)
}
\arguments{
\item{s}{Total number of successes.}

\item{n1}{Stage 1 sample size.}

\item{r1}{Stage 1 critical value (trial is stopped at stage 1 if the number of successes
is at most \code{r1}).}

\item{n}{Total sample size.}

\item{p0}{Response rate under the null hypothesis.}

\item{pvalue}{The desired p-value.}
}
\value{
Response probability.
}
\description{
Finds, for Simon-like designs, the value of response probability that
would yield a given p-value.
}
\details{
The solution is found using numerical search, with a precision of 0.000001.
The p-value is as defined by \emph{Koyama and Chen (2008)}.
}
\examples{
pquantile(21, 19, 4, 54, 0.2, 0.5)
}
\references{
Koyama, T. and Chen, H. Proper inference from Simon's two-stage designs.
\emph{Stat Med}, 2008, 27, 3145-3154.
}
\seealso{
\code{\link{pvaluek}} and \code{\link{pk}}.
}
\author{
Arsenio Nhacolo
}
