% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pwr.r
\name{Pwr}
\alias{Pwr}
\title{Power for single-stage parallel-group RCT.}
\usage{
Pwr(pc, pt, Nc, alp = 0.05)
}
\arguments{
\item{pc}{Response probability in control group.}

\item{pt}{Response probability in treatment group.}

\item{Nc}{Sample size per group.}

\item{alp}{Significance level (default: 0.05).}
}
\value{
Sample size for one group.
}
\description{
\code{Pwr} calculates the power in an equal-size group
two-arm randomized clinical trial with a binary response.
}
\examples{
Pwr(0.2,0.3,389,0.05)
}
\references{
Ahn, C., Heo, M. and Zhang, S. \emph{Sample Size Calculations for Clustered and Longitudinal Outcomes in Clinical Research}.
CRC Press, 2014.
}
\seealso{
\code{\link{Nct}}.
}
\author{
Arsenio Nhacolo
}
