% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbias.r
\name{sbias}
\alias{sbias}
\title{Bias of the sample proportion}
\usage{
sbias(n1, r1, n, p)
}
\arguments{
\item{n1}{Stage 1 sample size.}

\item{r1}{Stage 1 critical value (trial is stopped at stage 1 if the number of successes
is at most \code{r1}).}

\item{n}{Total sample size.}

\item{p}{True success probability.}
}
\value{
Bias.
}
\description{
Calculates bias due to using sample proportion as estimator of the true
response rate.
}
\details{
For fixed designs the sample propotion is an unbiased (maximum likelihood)
estimator of the response rate, but in group sequential designs (e.g., Simon's)
it is biased.
}
\examples{
sbias(19, 4, 54, 0.4)
}
\references{
Porcher, R. and Desseaux, K. What inference for two-stage phase II trials?
\emph{BMC Med Res Methodol},  2012, 12, 117.
}
\seealso{
\code{\link{sfms}} and \code{\link{pg}}.
}
\author{
Arsenio Nhacolo
}
