% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfms.r
\name{sfms}
\alias{sfms}
\title{Probability mass function of \code{(M, S)}}
\usage{
sfms(s, n1, r1, n, p, m = NA)
}
\arguments{
\item{s}{Total number of successes.}

\item{n1}{Stage 1 sample size.}

\item{r1}{Stage 1 critical value (trial is stopped at stage 1 if the number of successes
is at most \code{r1}).}

\item{n}{Total sample size.}

\item{p}{True success probability.}

\item{m}{Stage number (1 or 2). It is automatically determined based on \code{s} and \code{r1},
therefore it shouldn't be provided, unless there are reasons to do so.}
}
\value{
Density.
}
\description{
Probability mass function of \code{M} (stage) and \code{S} (number of successes).
}
\details{
Probability mass function of the statistic \code{(M, S)} for Simon-like
designs (allowing early stopping for futility only).
}
\examples{
sfms(21, 19, 4, 54, 0.4)
}
\references{
Jung, S.-H. and Kim, K. M. On the estimation of the binomial probability in
multistage clinical trials. \emph{Stat Med}, 2004, 23, 881-896.
}
\seealso{
\code{\link{sbias}} and \code{\link{pg}}.
}
\author{
Arsenio Nhacolo
}
