% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateEKOAD.r
\name{SimulateEKOAD}
\alias{SimulateEKOAD}
\title{Simulate single-arm binary endpoint two-stage adaptive designs.}
\usage{
SimulateEKOAD(replicates, dsgn, newpi1 = NULL, seed = NULL,
  deleteOld = TRUE)
}
\arguments{
\item{replicates}{Number of trials to be simulated.}

\item{dsgn}{Dataframe containing one of the designs in \code{\link{EKOADwn}}.}

\item{newpi1}{New response rate under the alternative hypothesis used to simulate trials. If \code{NULL} (default),
the one from the design is used.}

\item{seed}{The seed for random number generator. If \code{NULL} (default), no seed is set and , hence, results
are not reproducible.}

\item{deleteOld}{If \code{TRUE} (default), the simulation sub-directory is cleared before simulations start.}
}
\value{
Simulated trials are saved in the sub-directory ./SimulatedTrials.
}
\description{
\code{SimulateEKOAD} Simulate trials following designs similar to that of Englert and Kieser(2013)'s.
}
\details{
The original designs (like the ones in \code{\link{EKOptAdaptDesigns}}) must be pre-processed
using the function \code{\link{dsgnPrep}} to get extra information like the designs in \code{\link{EKOADwn}}.
}
\references{
Englert, S. and Kieser, M. Optimal adaptive two-stage designs for phase {II} cancer clinical trials.
\emph{Biometrical Journal}, 2013.
}
\seealso{
\code{\link{EKOptAdaptDesigns}}, \code{\link{EKOADwn}}.
}
\author{
Arsenio Nhacolo
}
