% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateSimonDsgn.r
\name{SimulateSimonDsgn}
\alias{SimulateSimonDsgn}
\title{Simon's designs data simulation}
\usage{
SimulateSimonDsgn(replicates, designParam, newp1 = NA, seed = NA,
  deleteOld = TRUE)
}
\arguments{
\item{replicates}{Number of trials to be generated.}

\item{designParam}{A dataframe containing Simon's optimal and minimax designs, as returned
by the function \code{\link{CalculateSimonDsgn}}.}

\item{newp1}{If \code{NA} (default) data are generated assuming the same response probability
under alternative hypothesis, \code{p1}, used to get the designs
(see \code{\link{CalculateSimonDsgn}}). One may provide different values of \code{newp1} if
there is interest in studying the effect of departure from the design's assumed \code{p1}.}

\item{seed}{Initial value (any integer) of random-number seed. It  is useful for creating
simulations that can be reproduced. The default is \code{NA}, meaning no reproducibility.}

\item{deleteOld}{If TRUE (default) the sub-directories \code{/Optimal/SimulatedTrials} and
\code{/Minimax/SimulatedTrials} are deleted, if they exist, before simulation starts. The
old data files are still replaced by the new ones even if \code{deleteOld} is set to
\code{FALSE}, but some old files remain in cases where the previous \code{replicates} was
greater that the current one.}
}
\value{
The function is not intended to return an R object, instead it creates files
(in CSV format) containing simulated trials data. See \emph{Details}. It also saves
in the current working directory the \code{designParam} argument (\emph{DesignParameters.csv}).
}
\description{
\code{SimulateSimonDsgn} simulates data from Simon's optimal and minimax designs.
}
\details{
The simulated trials are stored in the sub-directories \code{/Optimal/SimulatedTrials}
and \code{/Minimax/SimulatedTrials} for optimal and minimax designs, repectively, under the
current working directory. The sub-directories are automatically created. Individual trial
data are stored in a CSV file named \code{trial#}, where # is the replicate number.
}
\examples{
d <- CalculateSimonDsgn(0.2, 0.4, 0.05, 0.1)
SimulateSimonDsgn(100, d, seed = 1986)
}
\seealso{
\code{\link{CalculateSimonDsgn}}, \code{\link{SimulateSimonDsgnAdaptN}}
and \code{\link{AnalyzeSimonDsgn}}.
}
\author{
Arsenio Nhacolo
}
